# swagger.yaml
openapi: 3.0.0
info:
  title: User & Event API
  description: API for managing users and events
  version: 1.0.0
servers:
  - url: http://192.168.0.103:3000
    # - url: http://192.168.68.145:3000
    description: Local server
components:
  schemas:
...
    Error:
      type: object
      properties:
        status:
          type: string
          example: ERROR
        code:
          type: integer
          example: 400
        error:
          type: object
          properties:
            message:
              type: string
              example: 'Error message'

    # Success Response for User
...
    EventsListResponse:
      type: object
      properties:
        status:
          type: string
          example: OK
        code:
          type: integer
          example: 200
        data:
          type: array
          items:
            $ref: '#/components/schemas/Event'

    # Event Schema
    Event:
      type: object
      properties:
        event_id:
          type: integer
          example: 1
        event_name:
          type: string
          maxLength: 255
          example: 'Tech Conference 2025'
        event_date:
          type: string
          format: date
          example: '2025-05-15'
        city:
          type: string
          maxLength: 100
          example: 'San Francisco'
        description:
          type: string
          example: 'An exciting tech conference for developers'
        category:
          type: string
          maxLength: 100
          example: 'Technology'
        latitude:
          type: number
          format: float
          example: 37.7749
        longitude:
          type: number
          format: float
          example: -122.4194
    EventResponse:
      type: object
      properties:
        status:
          type: string
          example: OK
        code:
          type: integer
          example: 200
        data:
          $ref: '#/components/schemas/Event'

    # Request Body for Creating Event
    CreateEventRequest:
      type: object
      properties:
        event_name:
          type: string
          maxLength: 255
          example: 'Tech Conference 2025'
        event_date:
          type: string
          format: date
          example: '2025-05-15'
        city:
          type: string
          maxLength: 100
          example: 'San Francisco'
        description:
          type: string
          example: 'An exciting tech conference for developers'
        category:
          type: string
          maxLength: 100
          example: 'Technology'
        latitude:
          type: number
          format: float
          example: 37.7749
        longitude:
          type: number
          format: float
          example: -122.4194
...
    MyEventListRequest:
      type: object
      properties:
        userId:
          type: integer
          description: The ID of the user for whom to fetch RSVPed events.
          example: 123
        isUpcoming:
          type: boolean
          description: Flag to filter for upcoming events (e.g., `true` for upcoming events).
          default: false
          example: true
        isPast:
          type: boolean
          description: Flag to filter for past events (e.g., `true` for past events).
          default: false
          example: false
        isToday:
          type: boolean
          description: Flag to filter for today's events (e.g., `true` for today's events).
          default: false
          example: false
      required:
        - userId
paths:
...
  /api/event/myEvent/list:
    post:
      operationId: 'getMyEvents'
      tags:
        - Events
      summary: Get a list of events a user has RSVPed to
      description: This endpoint returns a list of events that the user has RSVPed to, based on the isUpcoming, isPast, and isToday flags.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MyEventListRequest'
      responses:
        '200':
          description: A list of events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventsListResponse'
        '400':
          description: Error retrieving events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/event/{id}:
    get:
      operationId: 'getEventById'
      summary: Get event by ID
      description: Retrieve an event by its unique ID
      tags:
        - Events
      parameters:
        - in: path
          name: id
          required: true
          description: The ID of the event
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: An event object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      operationId: 'updateEvent'
      summary: Update event by ID
      description: Update the event details for a given event ID
      tags:
        - Events
      parameters:
        - in: path
          name: id
          required: true
          description: The ID of the event to update
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '200':
          description: Event updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          description: Error updating event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: 'deleteEvent'
      summary: Delete event by ID
      description: Delete the event by its unique ID
      tags:
        - Events
      parameters:
        - in: path
          name: id
          required: true
          description: The ID of the event to delete
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Event deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          description: Error deleting event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...